// src/types/homepage.about.types.ts
export interface HomePageAboutRequest{
    leftTitle: string;
    leftDescription: string;
    rightTitle: string;
    rightDescription: string;
}

export interface HomePageAboutResponse{
    id:number;
    leftTitle: string;
    leftDescription: string;
    rightTitle: string;
    rightDescription: string;
}